changed routes to call function in controller file instead of anonymous callback. updated controller files to contain data that was in anon cbs in routes
