+++ views/all_articles.handlebars
@@ -1,6 +1,6 @@
 <nav class="navbar">
     <a class="navbar-brand" href="/">reddit r/WorldNews Scraper</a>
-    <button class="btn btn-danger" id="scrape-btn" type="submit">Scrape New Headlines</button>
+    <button class="btn btn-danger" id="scrape-btn" type="button">Scrape New Headlines</button>
 
 </nav>
