+++ views/one_article.handlebars
@@ -60,7 +60,7 @@
                     <p>{{this.title}}</p>
                     <p>{{this.text}}</p>
 
-                    <button type="submit" class="btn btn-primary mb-2 delete-note" data-id={{this._id}}>Delete Note</button>
+                    <button type="button" class="btn btn-primary mb-2 delete-note" data-id={{this._id}}>Delete Note</button>
 
                 </div>
